Assistant checkpoint: Update TaskCard component design
Assistant generated file changes:
- src/components/TaskList.tsx: Update TaskCard design to match reference

---

User prompt:

please make the task element like this  only these many fields will be shown. as you can see there is a checkbox to select the element then there is an icon of 6 dots. please create this icon and after that there is another checkbox of round style this one is to indicate that the task is completed.